+++ scripts/Changelog.md
@@ -16,6 +16,13 @@
 ### 1.0.3
 1. Added update-script command
 2. Ask if you want to install pre-request packages first
-3. Shut down server before update if running
+3. Shutdown server before update if running
 4. Send messages to online players about checking for server updates, when its shutting down, and when its restarting
 5. Wait 10 seconds before shutdown/restart to allow players time to react to shutdown/restart messages
+
+### 1.1.0
+1. Added option to import .mcworld files from ./content/worlds folder
+2. Added option to import .mcpack files from ./content/addons folder
+   - This will only update already installed addons
+3. Stop server before backup
+4. Cleaned up some repetitive code